Adjusts on authentication route
